create series and arch arguments for integration tests
