+++ tox.ini
@@ -57,10 +57,10 @@ deps =
 commands =
     coverage run --source={[vars]src_path} \
                  -m pytest \
-                 --ignore={[vars]tst_path}integration \
                  --tb native \
                  -v \
                  -s \
+                 {[vars]tst_path}unit
                  {posargs}
     coverage report
 
@@ -82,6 +82,6 @@ commands =
     pytest -v \
            -s \
            --tb native \
-           --ignore={[vars]tst_path}unit \
            --log-cli-level=INFO \
+           {[vars]tst_path}integration \
            {posargs}
